+++ documentation/changelog.md
@@ -37,6 +37,8 @@ description: List of all the changes and current versions.
 + Added Custom Interaction Text
 + Added Custom Interaction Blips
 + Added Custom Interaction Functionality
++ Added Blip Controller
++ Added Global Blip Propogation
 ```
 
 ## 1.5.0 - January 28-31, 2021